requirements.txt
asgiref==3.7.2
Django==4.2.2
djangorestframework==3.14.0
Pillow==9.5.0
psycopg==3.1.9
pyscopg2==66.0.2
python-dotenv==1.0.0
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.7.0
